Refactor model.
