+++ client_model.go
